+++ Import/Dynamic/ComBinding.cs
@@ -6,7 +6,7 @@ namespace Kaenx.DataContext.Import.Dynamic
 {
     public class ComBinding
     {
-        public int ComId { get; set; }
+        public long ComId { get; set; }
         public List<ParamCondition> Conditions { get; set; } = new List<ParamCondition> ();
     }
 }
